fs_partition_resize: perform reboot for journal only if initramfs
The problem with next mount after adding a journal exists only if the initramfs performs an fsck before mounting the rootfs. It migrates the temporary /.journal inode into the hidden inode 8, and for whatever reason, this causes a failing mount, which is solved only with a forced fsck with forced repair, otherwise requires another power cycle, before fsck fixes it without "-y".

If there is no initramfs (or it does not fsck), systemd-fsck runs instead, but it cannot migrate /.journal to the hidden inode 8, since the rootfs is (R/O) mounted already. It hence also does not cause the mount issue, no reboot and no forced fsck is needed. This is the reason this issue was not recognised before.

There is still a little benefit when performing the reboot regardless, since the journal is pratically not in use before. It needs to be freshly mounted to become active, the R/W-remount is not sufficient. But a reboot is anyway often needed after first run setup, so that should be fine, not slowing down first run setup.

Add changelog entry about the issue.

+++ rootfs/var/lib/dietpi/services/fs_partition_resize.sh
@@ -151,12 +151,17 @@
 				resize2fs "$ROOT_DEV" || REBOOT='since the root filesystem resize failed' # https://github.com/MichaIng/DietPi/issues/6149
 				if [[ $ROOT_FSTYPE == 'ext'[34] ]] && ! tune2fs -l "$ROOT_DEV" | grep -q 'has_journal'
 				then
-					echo '[ INFO ] Adding filesystem journal and performing a reboot with forced fsck'
-					REBOOT='to apply the new root filesystem journal'
-					echo -e '#!/bin/dash\ntune2fs -c 0 '"'$ROOT_DEV' && rm /var/lib/dietpi/postboot.d/dietpi-reset_max_mount_count" > /var/lib/dietpi/postboot.d/dietpi-reset_max_mount_count
-					tune2fs -O 'has_journal' -c 1 -C 2 "$ROOT_DEV"
-					sync
-					sleep 1
+					echo '[ INFO ] Adding root filesystem journal'
+					if [[ -e '/run/initramfs/fsck-root' ]]
+					then
+						REBOOT=' with forced fsck to apply the new root filesystem journal'
+						echo -e '#!/bin/dash\ntune2fs -c 0 '"'$ROOT_DEV' && rm /var/lib/dietpi/postboot.d/dietpi-reset_max_mount_count" > /var/lib/dietpi/postboot.d/dietpi-reset_max_mount_count
+						tune2fs -O 'has_journal' -c 1 -C 2 "$ROOT_DEV"
+						sync
+						sleep 1
+					else
+						tune2fs -O 'has_journal' "$ROOT_DEV"
+					fi
 				fi
 			else
 				echo '[ INFO ] Skipping root filesystem expansion since detected root partition device node does not exist, assuming container system'
